PlaceHandler: fetch place prior to delete
We return the deleted place after delete so it is necessary
to fetch it first. This introduces a race condition to be
later solved by Dao transactions.

Fixes: 751234f848f5 ("PlaceService: new class PlaceService")
Fixes: b49a12fae9dc ("PlaceHandler: new class")

+++ server/src/service/PlaceService.h
@@ -39,8 +39,11 @@ public:
 		return m_placeDao->update(place);
 	}
 
-	bool remove(const Place &place)
+	bool remove(Place &place)
 	{
+		if (!m_placeDao->fetch(place))
+			return false;
+
 		return m_placeDao->remove(place);
 	}
